a very hacky version of 'using threads'
This uses std::thread::spawn to handle each new stream in the TCP
listener in a separate OS thread. It works! One funky bit is that for
lifetime reasons and thread-safety reasons everything relevant to the
new thread has to be moved within the closure that's used to start it,
so I needed to reconfigure my `main` function a little bit but otherwise
not much needed to change actually.

Still remaining: right now if you had a very long-running response and
you killed the main process we could have a zombie. I need to add
something to intercept the ctrl-c interrupt and kill all the child
processes before the parent process exits.