add thread pool and interrupt handling
This adds proper multi-threading to Bippy.

I based the implementation on the code in chapter 20 of the Rust book
but changed things to make the TCP listener handling code non-blocking.
This allows the webserver to correctly handle ctrl-c interrupts and
thereby clean up after itself, rather than leaving orphan threads around
for the OS to clean up later.